pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "API for the Donut-Diaries"
authors = ["Brian Kim <bryo.kim1@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
bullmq = "^2.3.3"
pyjwt = "^2.8.0"
python-decouple = "^3.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
